backend/server.py
from fastapi import FastAPI, HTTPException, UploadFile, File
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
from motor.motor_asyncio import AsyncIOMotorClient
from pydantic import BaseModel
import os
from dotenv import load_dotenv
import uvicorn
from typing import Optional, List
import logging

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(title="LotayaAI API", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# MongoDB connection
MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017/lotayaai")
client = None
db = None

@app.on_event("startup")
async def startup_event():
    global client, db
    try:
        client = AsyncIOMotorClient(MONGO_URL)
        db = client.get_database()
        logger.info("Connected to MongoDB successfully")
    except Exception as e:
        logger.error(f"Failed to connect to MongoDB: {e}")

@app.on_event("shutdown")
async def shutdown_event():
    if client:
        client.close()
        logger.info("Disconnected from MongoDB")

# Pydantic models
class ImageGenerationRequest(BaseModel):
    prompt: str
    model: str = "groq"  # groq, xai, gemini
    style: Optional[str] = None
    size: Optional[str] = "1024x1024"

class VideoGenerationRequest(BaseModel):
    prompt: str
    model: str = "runway"  # runway, kling, veo3, sora
    duration: Optional[int] = 10

class TextToVideoRequest(BaseModel):
    script: str
    model: str = "runway"
    style: Optional[str] = None

# API Routes
@app.get("/api/health")
async def health_check():
    return {"status": "healthy", "message": "LotayaAI API is running"}

@app.get("/api/models")
async def get_available_models():
    return {
        "image_models": ["groq", "xai", "gemini"],
        "video_models": ["runway", "kling", "veo3", "sora", "seedance", "hailuo"],
        "effects": ["ai_hug", "ai_kissing", "french_kiss", "decapitate", "eye_pop"]
    }

@app.post("/api/generate/image")
async def generate_image(request: ImageGenerationRequest):
    try:
        # TODO: Implement actual AI image generation
        # For now, return a placeholder response
        return {
            "success": True,
            "message": "Image generation initiated",
            "model_used": request.model,
            "prompt": request.prompt,
            "image_url": "placeholder_image_url",
            "generation_id": "temp_id_123"
        }
    except Exception as e:
        logger.error(f"Image generation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/generate/video")
async def generate_video(request: VideoGenerationRequest):
    try:
        # TODO: Implement actual AI video generation
        return {
            "success": True,
            "message": "Video generation initiated",
            "model_used": request.model,
            "prompt": request.prompt,
            "video_url": "placeholder_video_url",
            "generation_id": "temp_video_id_123"
        }
    except Exception as e:
        logger.error(f"Video generation error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/convert/text-to-video")
async def text_to_video(request: TextToVideoRequest):
    try:
        # TODO: Implement text to video conversion
        return {
            "success": True,
            "message": "Text to video conversion initiated",
            "script": request.script,
            "model_used": request.model,
            "video_url": "placeholder_video_url",
            "conversion_id": "temp_conversion_id_123"
        }
    except Exception as e:
        logger.error(f"Text to video conversion error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/api/generations/{generation_id}")
async def get_generation_status(generation_id: str):
    try:
        # TODO: Implement actual status checking
        return {
            "generation_id": generation_id,
            "status": "completed",
            "progress": 100,
            "result_url": "placeholder_result_url"
        }
    except Exception as e:
        logger.error(f"Status check error: {e}")
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8001)